Create Selection Class
The new Selection class contains methods and properties used for
the validation of the options provided to the component.

Consider that it is missing some important features that haven't
been tested design wise.A few of them are the following:
- A style class that changes the emphasized color to grey used
only when the attribute disabled exists.
- A spinner icon that appears only when the options provided
depend on a async process.

See: #31