Create Input Collection Class
This new track class for form fields is based off a rather similar
set of methods and properties used in `NewWebsiteModal.js` .  I noticed
that it would be nice to have a global solution rather than be repeating
the same methods all over the different modal files.

The collection class can retrieve, add or remove existing entries within
the main array of objects.

There are two inner classes within that are used to designed either a
native or a custom element field. These two classes share the same shape
and only vary in the methods that they define.